+++ liqwid-plutarch-extra/src/Plutarch/Extra/AssetClass.hs
@@ -53,7 +53,6 @@ import Optics.Lens (A_Lens)
 import Optics.Optic (Is, (%%))
 import Optics.Setter (set)
 import Optics.TH (makeFieldLabelsNoPrefix)
-import Plutarch.Api.V1 (PCurrencySymbol, PTokenName)
 import Plutarch.DataRepr (PDataFields)
 import Plutarch.Extra.Applicative (ppure)
 import Plutarch.Extra.IsData (
@@ -62,6 +61,7 @@ import Plutarch.Extra.IsData (
  )
 import Plutarch.Extra.Record (mkRecordConstr, (.&), (.=))
 import Plutarch.Extra.Tagged (PTagged)
+import Plutarch.LedgerApi (PCurrencySymbol, PTokenName)
 import Plutarch.Lift (
   PConstantDecl,
   PUnsafeLiftDecl (PLifted),