+++ liqwid-plutarch-extra/src/Plutarch/Extra/Fixed.hs
@@ -14,10 +14,13 @@ import Control.Composition (on, (.*))
 import Data.Bifunctor (first)
 import Data.Proxy (Proxy (Proxy))
 import GHC.TypeLits (KnownNat, Nat, natVal)
-import Plutarch.Api.V1 (AmountGuarantees (NonZero), PValue)
-import Plutarch.Api.V1.Value qualified as Value
-import Plutarch.Api.V2 (KeyGuarantees (Sorted))
 import Plutarch.Extra.Function (pflip)
+import Plutarch.LedgerApi (
+  AmountGuarantees (NonZero),
+  KeyGuarantees (Sorted),
+  PValue,
+ )
+import Plutarch.LedgerApi.Value qualified as Value
 import Plutarch.Num (PNum (pfromInteger, (#*)))
 import Plutarch.Numeric.Additive qualified as A (
   AdditiveMonoid (zero),