+++ liqwid-plutarch-extra/src/Plutarch/Extra/Functor.hs
@@ -13,11 +13,14 @@ module Plutarch.Extra.Functor (
 ) where
 
 import Data.Kind (Constraint)
-import Plutarch.Api.V1.AssocMap (KeyGuarantees, PMap (PMap))
-import Plutarch.Api.V1.Maybe (PMaybeData (PDJust, PDNothing))
 import Plutarch.Builtin (ppairDataBuiltin)
 import Plutarch.Extra.Boring (PBoring (pboring))
 import Plutarch.Extra.Function (pconst, pidentity)
+import Plutarch.LedgerApi (
+  KeyGuarantees,
+  PMap (PMap),
+  PMaybeData (PDJust, PDNothing),
+ )
 import Plutarch.Lift (PUnsafeLiftDecl)
 
 {- | Describes the entire category of Plutarch types, with arrows being Plutarch