+++ liqwid-plutarch-extra/src/Plutarch/Extra/MultiSig.hs
@@ -22,13 +22,13 @@ import GHC.Records (HasField)
 import Optics.Getter (A_Getter, to)
 import Optics.Label (LabelOptic (labelOptic))
 import Optics.Traversal (A_Traversal, traversalVL)
-import Plutarch.Api.V2 (PPubKeyHash)
 import Plutarch.DataRepr (
   DerivePConstantViaData (DerivePConstantViaData),
   PDataFields,
  )
 import Plutarch.Extra.Field (pletAllC)
 import Plutarch.Extra.Function (pflip)
+import Plutarch.LedgerApi (PPubKeyHash)
 import Plutarch.Lift (PConstantDecl, PLifted, PUnsafeLiftDecl)
 import PlutusLedgerApi.V1.Crypto (PubKeyHash)
 import PlutusTx qualified (makeLift, unstableMakeIsData)