+++ liqwid-plutarch-extra/src/Plutarch/Extra/Rational.hs
@@ -1,4 +1,3 @@
-
 {-# LANGUAGE PolyKinds #-}
 {-# LANGUAGE RankNTypes #-}
 
@@ -255,7 +254,7 @@ ptoPositiveCases n contNeg contPos =
     (n #<= 0)
     ( pif
         (n #== 0)
-        (ptraceError "ptoPositiveCases with 0")
+        (ptraceInfoError "ptoPositiveCases with 0")
         -- The PPositive constructor is not exported, so we need coercion
         (contNeg (punsafeCoerce $ -n))
     )