+++ liqwid-plutarch-extra/test/ValueProp.hs
@@ -4,9 +4,9 @@
 module ValueProp (tests) where
 
 import Data.Maybe (catMaybes)
-import Plutarch.Api.V2 (PMap (PMap))
 import Plutarch.Builtin (ppairDataBuiltin)
 import Plutarch.Extra.Value (phasOnlyOneTokenOfCurrencySymbol, pvalue)
+import Plutarch.LedgerApi (PMap (PMap))
 import Plutarch.Test.QuickCheck (fromPFun)
 import Test.QuickCheck (
   Property,