+++ plutarch-unit/src/Plutarch/Test/Unit.hs
@@ -26,19 +26,21 @@ module Plutarch.Test.Unit (
 
 import Data.Tagged (Tagged (Tagged))
 import Data.Text (Text)
-import Plutarch.Api.V2 (PMintingPolicy, PValidator)
 import Plutarch.Builtin (pforgetData)
 import Plutarch.Evaluate (EvalError, evalScript)
 import Plutarch.Extra.Compile (mustCompile)
+import Plutarch.LedgerApi (PScriptContext)
 import Plutarch.Lift (PUnsafeLiftDecl (PLifted))
 import Plutarch.Prelude (
   PData,
+  POpaque,
   S,
   Term,
   Type,
   pconstant,
   pconstantData,
   (#),
+  (:-->),
  )
 import Plutarch.Script (Script)
 import PlutusLedgerApi.V2 (ScriptContext)
@@ -89,7 +91,7 @@ mintingPolicySucceedsWith ::
   forall (redeemer :: S -> Type).
   (PUnsafeLiftDecl redeemer, ToData (PLifted redeemer)) =>
   String ->
-  (forall (s :: S). Term s PMintingPolicy) ->
+  (forall (s :: S). Term s (PData :--> PScriptContext :--> POpaque)) ->
   PLifted redeemer ->
   ScriptContext ->
   TestTree
@@ -104,7 +106,7 @@ mintingPolicyFailsWith ::
   forall (redeemer :: S -> Type).
   (PUnsafeLiftDecl redeemer, ToData (PLifted redeemer)) =>
   String ->
-  (forall (s :: S). Term s PMintingPolicy) ->
+  (forall (s :: S). Term s (PData :--> PScriptContext :--> POpaque)) ->
   PLifted redeemer ->
   ScriptContext ->
   TestTree
@@ -123,7 +125,7 @@ validatorSucceedsWith ::
   , ToData (PLifted redeemer)
   ) =>
   String ->
-  (forall (s :: S). Term s PValidator) ->
+  (forall (s :: S). Term s (PData :--> PData :--> PScriptContext :--> POpaque)) ->
   PLifted datum ->
   PLifted redeemer ->
   ScriptContext ->
@@ -143,7 +145,7 @@ validatorFailsWith ::
   , ToData (PLifted redeemer)
   ) =>
   String ->
-  (forall (s :: S). Term s PValidator) ->
+  (forall (s :: S). Term s (PData :--> PData :--> PScriptContext :--> POpaque)) ->
   PLifted datum ->
   PLifted redeemer ->
   ScriptContext ->
@@ -161,7 +163,7 @@ testValidator ::
   ) =>
   (String -> Script -> TestTree) ->
   String ->
-  (forall (s :: S). Term s PValidator) ->
+  (forall (s :: S). Term s (PData :--> PData :--> PScriptContext :--> POpaque)) ->
   PLifted datum ->
   PLifted redeemer ->
   ScriptContext ->
@@ -187,7 +189,7 @@ testMP ::
   (PUnsafeLiftDecl redeemer, ToData (PLifted redeemer)) =>
   (String -> Script -> TestTree) ->
   String ->
-  (forall (s :: S). Term s PMintingPolicy) ->
+  (forall (s :: S). Term s (PData :--> PScriptContext :--> POpaque)) ->
   PLifted redeemer ->
   ScriptContext ->
   TestTree